Changed snapshot reaping to use received properties, not *-snapshot-limit
This addresses the problem of reaping snapshots on the destination
replica.  Previously I tried to do this by not replicating properties
(avoiding zfs send -R), and introducing a new property per tier,
*tier*-snapshot-limit.  This was a mistake.  So having embraced the
fact that properties will in fact be replicated, we detect this on the
destination, as the source is received.  For such properties, no new
snapshots or replication is performed.  But old snapshots are reaped
anyway.  So running zbackup on a replica destination host simply
serves to cleanup old snapshots, safely.