zsnap: --warnondestroyfailure option.
Also, zbackup --zsnap-options to pass options to zsnap.

Making destory failures a warning rather than a fatal error is useful in
case replications are taking a long time to run, so that an old snapshot which
is due for deletion may still be being replicated.  This option therefore allows
to continue with a warning.  The snapshot is expected to be deleted later, once
replication has finished.