Bug Fix for Python 2.x Related To Prior PR
The last PR I merged had a regression with Python 2.x because a `str` in Python 2 is also a `bytes`. This Python 2 quirk causes the decode('utf-8') block to always execute. This patch detects the Python version and only calls decode('utf-8') if its not python 2.

I discovered this whilst getting zfslib working on Python 2.7.18. Sorry I did not catch this bug earlier.